Conflicts:
	_class/UserImpl.php
	_dart/pcre_syntax_checker.dart
	_dart/src/json_response.dart
	_dart/src/site_class_page.dart
	_interface/BackendSingletonContainer.php
	_interface/JSONResponse.php
	_test/UserImplTest.php

+++ _dart/src/core_status_bar.dart
@@ -42,7 +42,7 @@ class SavingBar{
           _statusBar.remove();
         }
       });
-      _updatePercentage(1);
+      _updatePercentage(1.toDouble());
       return;
     }
     if(_endedJobs.length == 0 && _runningJobs.length == 1){
